+++ src/dreamstate/core/observing/startServiceObserving.ts
@@ -2,7 +2,9 @@ import { CONTEXT_SERVICES_REFERENCES, CONTEXT_SERVICES_REGISTRY } from "@/dreams
 import { TDreamstateService } from "@/dreamstate/types";
 
 /**
- * todo;
+ * Start observing of service and trigger related lifecycle methods.
+ * Count providers references so it will work with few different providers
+ * and force 'onProvisionStarted' to be used only once.
  */
 export function startServiceObserving(
   Service: TDreamstateService